video-watcher/requirements.txt
python-telegram-bot
watchdog
paho-mqtt